+++ RestaurantManagement/Models/Kho.cs
@@ -11,15 +11,15 @@ namespace QuanLyNhaHang.Models
     {
         private string _TenSanPham;
         public string TenSanPham { get => _TenSanPham; set { _TenSanPham = value; OnPropertyChanged(); }}
-        private int _TonDu;
-        public int TonDu { get => _TonDu; set { _TonDu = value; OnPropertyChanged(); } }
+        private float _TonDu;
+        public float TonDu { get => _TonDu; set { _TonDu = value; OnPropertyChanged(); } }
         private string _DonVi;
         public string DonVi { get => _DonVi; set { _DonVi = value; OnPropertyChanged(); } }
         private string _DonGia;
         public string DonGia { get => _DonGia; set { _DonGia = value; OnPropertyChanged(); } }
 
 
-        public Kho(string ten, int tondu, string donvi, string dongia)
+        public Kho(string ten, float tondu, string donvi, string dongia)
         {
             TenSanPham = ten;
             TonDu = tondu;
